requirements.txt
# HIMARI Opus1 - Requirements
# Install with: pip install -r requirements.txt

# Flink
apache-flink>=1.18.0
kafka-python>=2.0.2
redis>=4.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
neo4j>=5.0.0
numpy
pandas
pytest
pytest-asyncio
requests
python-dotenv
pyyaml
sops
